Move LicenseInfo to com.hazelcast.internal.management
